fix(swa): apply_rotary_emb_3d must return the input dtype
build_3d_rope emits fp32 cos/sin (correct for angle precision), so rotating a
bf16 q/k promoted the result to fp32 with no cast back. v is not rotated and
stays bf16, so SDPA rejected the mixed dtypes (query float != value bfloat16)
and the flash path silently upcast every block. The adjacent _qk_norm already
casts with .to(x.dtype); the rotation was the missing case.

Do the rotation math in the promoted (fp32) dtype but cast the result back to
the input dtype. Verified on H100: bf16/fp16/fp32 inputs all return the same
dtype, cos-sim 1.0 vs the fp32 reference.

+++ src/miniworld_engine/modules/swa_atom_attention.py
@@ -270,7 +270,11 @@ def apply_rotary_emb_3d(
     cos = cos.unsqueeze(2).repeat(1, 1, 1, 2)  # [N, S, 1, 2*half]
     sin = sin.unsqueeze(2).repeat(1, 1, 1, 2)
     x_rot = x[..., :ro_dim]
-    x_rot = x_rot * cos + _rotate_half(x_rot) * sin
+    # cos/sin are fp32 (angle precision), which promotes a bf16 x to fp32. Do the
+    # rotation math in the promoted dtype but cast back to the input dtype, so q/k
+    # keep the same dtype as the (unrotated) v — otherwise SDPA rejects the mixed
+    # dtypes and flash silently upcasts. Mirrors _qk_norm's `.to(x.dtype)`.
+    x_rot = (x_rot * cos + _rotate_half(x_rot) * sin).to(x.dtype)
     return torch.cat([x_rot, x[..., ro_dim:]], dim=-1)
